Don't cure, immunize: ensure boost does not receive exceptions from modify()
Current boost will delete an object from a multi_index_container if when
modifying that object, the functor throws an exception. This breaks the
undo infrastructure when it tries to undo the failed change, but the
object is missing. To prevent this, we catch the exception before it
reaches boost.